--- components/code-block/README.md
@@ -1,13 +1,0 @@
-# Component: Code Block
-
-**Code block component.**
-
-## Get Started
-
-```sh
-npm i "code-block"
-```
-
-```
-import { CodeBlock } from "code-block";
-```
