--- components/code-block/src/code-block/code-block.types.ts
@@ -1,5 +1,0 @@
-export type CodeBlockProps = {
-  language: string;
-  code: string;
-  isCopyable?: boolean;
-};
