--- components/code-block/src/code-block/index.ts
@@ -1,3 +1,0 @@
-export { CodeBlock } from './code-block';
-export { useCodeBlockStyles } from './code-block.styles';
-export type { CodeBlockProps } from './code-block.types';
