--- components/code-block/src/index.ts
@@ -1,2 +1,0 @@
-export type { CodeBlockProps } from './code-block';
-export { CodeBlock, useCodeBlockStyles } from './code-block';
