+++ components/code-snippet/.eslintignore
