+++ components/code-snippet/.npmignore
