+++ components/code-snippet/README.md
@@ -1,0 +1,13 @@
+# Component: Code Snippet
+
+**Code Snippet component for simple code examples.**
+
+## Get Started
+
+```sh
+npm i @microsoft/arbutus.code-snippet
+```
+
+```jsx
+import { CodeSnippet } from "@microsoft/arbutus.code-snippet";
+```
