+++ components/code-snippet/package.json
@@ -1,8 +1,7 @@
 {
-  "name": "@microsoft/arbutus.code-block",
-  "private": true,
-  "version": "0.0.1",
-  "description": "Code Block component.",
+  "name": "@microsoft/arbutus.code-snippet",
+  "version": "0.0.1-alpha.0",
+  "description": "Code Snippet component for simple code examples.",
   "main": "lib/index.js",
   "module": "lib/index.js",
   "typings": "lib/index.d.ts",
