+++ components/code-snippet/src/code-snippet/code-snippet.types.ts
@@ -1,0 +1,11 @@
+export type CodeSnippetProps = {
+  /**
+   * The language of the code snippet.
+   */
+  language: string;
+
+  /**
+   * The code snippet to display.
+   */
+  code: string;
+};
