+++ components/code-snippet/src/code-snippet/index.ts
@@ -1,0 +1,3 @@
+export { CodeSnippet } from './code-snippet';
+export { useCodeSnippetStyles } from './code-snippet.styles';
+export type { CodeSnippetProps } from './code-snippet.types';
