+++ components/code-snippet/src/index.ts
@@ -1,0 +1,2 @@
+export type { CodeSnippetProps } from './code-snippet';
+export { CodeSnippet, useCodeSnippetStyles } from './code-snippet';
