+++ components/code-snippet/tsconfig.json
